requirements.txt
# Core ML
xgboost>=2.0.0
lightgbm>=4.3.0
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
joblib>=1.3.0

# Polymarket SDK
py-clob-client>=0.18.0
web3>=6.15.0

# LLM APIs
anthropic>=0.28.0
openai>=1.30.0

# Data
requests>=2.31.0
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0
